requirements.txt
Flask==2.3.2
PyJWT==2.8.0
Werkzeug>=2.3.3
python-dotenv==1.0.0